Add python bindings for Convex/Mesh GetConvexHull()
This includes the renaming from convex_hull() to GetConvexHull() to better
conform to the style guide.

Note: as this makes Shape dependent on PolygonSurfaceMesh, the bindings
of the mesh types have been moved into their own .cc file so that they
can be evaluated before the shapes.
