+++ bindings/pydrake/geometry/geometry_py_common.cc
@@ -422,6 +422,8 @@ void DoScalarIndependentDefinitions(py::module m) {
         .def("filename", &Convex::filename, doc.Convex.filename.doc)
         .def("extension", &Convex::extension, doc.Convex.extension.doc)
         .def("scale", &Convex::scale, doc.Convex.scale.doc)
+        .def("GetConvexHull", &Convex::GetConvexHull,
+            doc.Convex.GetConvexHull.doc)
         .def(py::pickle(
             [](const Convex& self) {
               return std::make_pair(self.filename(), self.scale());
@@ -474,6 +476,7 @@ void DoScalarIndependentDefinitions(py::module m) {
         .def("filename", &Mesh::filename, doc.Mesh.filename.doc)
         .def("extension", &Mesh::extension, doc.Mesh.extension.doc)
         .def("scale", &Mesh::scale, doc.Mesh.scale.doc)
+        .def("GetConvexHull", &Mesh::GetConvexHull, doc.Mesh.GetConvexHull.doc)
         .def(py::pickle(
             [](const Mesh& self) {
               return std::make_pair(self.filename(), self.scale());
