+++ bindings/pydrake/geometry/test/common_test.py
@@ -443,6 +443,11 @@ class TestGeometryCore(unittest.TestCase):
         self.assertIn(junk_path, convex.filename())
         self.assertEqual(".ext", convex.extension())
         self.assertEqual(convex.scale(), 1.0)
+        with self.assertRaisesRegex(RuntimeError,
+                                    "MakeConvexHull only applies to"):
+            # We just need evidence that it invokes convex hull machinery; the
+            # exception for a bad extension suffices.
+            convex.GetConvexHull()
         assert_pickle(
             self, convex, lambda shape: [shape.filename(), shape.scale()])
 
@@ -471,6 +476,11 @@ class TestGeometryCore(unittest.TestCase):
         self.assertIn(junk_path, mesh.filename())
         self.assertEqual(".ext", mesh.extension())
         self.assertEqual(mesh.scale(), 1.0)
+        with self.assertRaisesRegex(RuntimeError,
+                                    "MakeConvexHull only applies to"):
+            # We just need evidence that it invokes convex hull machinery; the
+            # exception for a bad extension suffices.
+            mesh.GetConvexHull()
         assert_pickle(
             self, mesh, lambda shape: [shape.filename(), shape.scale()])
